move App to components
