+++ packages/styleguide/src/lib/styleguide.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './../App';
+import App from './../components/App';
 
 export * from '../components';
 export * from '../utils';
