+++ packages/styleguide/src/utils/index.js
@@ -1,2 +1,1 @@
 export { default as md } from './md';
-export { default as ComponentDocs } from './ComponentDocs';
